Fix piece bounds check
This fixes the I piece bounds check on the left by moving the origin
point for BlockGrid to the left by one.  This broke pretty much
everything else, so I also cleaned up the other block to field
translations to all use the same method.  It's still a mess, but it's
less of one now.