Clear Left/Right shift on next block
Fix a bug that persisted the button held event after a black had been
placed even when left or right was no longer being held.

A side effect of this is that even if left or right is being held, the
shift state is reset.  The button will have to be released and pressed
again to enter the shift state again.