Implement detecting T-Spins
T-Spin detection is implemented, however there is still an edge case for
a false positive.  If a rotation is attempted, but cannot be completed
due to a collision, the check for a T-Spin will be enabled.  This means
if a T block was dropped down a deep 2-column void where the last row
only has a single block empty, and the player attempts to rotate when
the block is at the bottom, the T-Spin check will occur and falsely
detect a sucessful T-Spin even though the block did not rotate.

Care need to be taken when fixing this edge case to not invalidate
proper T-Spins if the player tries and fails to rotate a second time
after a successfull T-Spin.

Example board setup that will trigger this false positive:

    XXXXX  XXX
    XXXXX  XXX
    XXXXX  XXX
    XXXXX  XXX
    XXXXX -XXX
    XXXXX--XXX
    XXXXXX-XXX