Implement Pausing
Pressing start in single player will now pause the game.  In two player
mode, both players must press start at the same time (this is not a
frame-perfect input).

When paused no input other than checking for the start button and no
game logic is processed.  The screen will also be turned black and
white.

I might want to do a palette swap instead of using the BW mode on the
PPU Mask, but that'll be later.