Add a practice menu; generalize the option menu code
The code behind the option menu has been abstracted a bit so it can be
used for other menus.  This menu was already pretty generic, but these
changes isolate the menu logic from the data for the menu.

The new practice menu will eventually load up a playfield and start a
single player game, similar to how the debug field worked.  There are
some practice boards already defined, however the functionality to load
up these boards is not implemented yet.

The option on the options menu to clear the high scores table was
intentionally broken.  It's being changed to a button instead of a hacky
boolean thing.  Additional changes will be needed to rewrite the hacky
boolean crap into a proper button action.  This button should also have
a confirmation before actually erasing data.