Finish re-implementing clearing save ram
The button on the options menu to clear all of save ram now asks for
confirmation before clearing anything.  This confirmation screen uses
the basic menu code instead of some specific crap just for this.

Additionally, a notification is displayed for two seconds after clearing
the ram before loading up the previous menu.