Add the FamiStudio audio driver
This required moving around some RAM stuff. Most notably the high scores
screen data is now in save ram.  Some other things were moved to zero
page, and the palette buffer has been moved to stack space.

The audio driver is in its own bank starting at $C000 (same as the utils
bank).  The exported funcions have been wrapped with some routines in
the fixed bank to handle the bank switching.  The first bank at $8000 is
also swapped out in these routines.  Eventually, the music data will be
in there instead of the same bank as the driver.

Only the test MIDI has been included so far, as well as some initial
sound effects.  The music starts on the main menu and isn't ever
switched or stopped.  The main menu is the only screen with any sound
effects currently.