Fix a bunch of timing issues
Fixed a lot of timing issues that arose once the audio driver was added
to the NMI.  This mainly consists of adding a PpuMask variable that is
written to the PPU register during NMI similar to the PpuControl
variable.

Also added a PPU write buffer that can be filled outside of NMI and
will be drawn during NMI.  This draw happens after the trampoline and
before the call to the sound driver update routine.