Start implementing VS mode garbage attacks
Players can now send garbage lines to each other upon clearing lines.
The number of lines sent is deterimned by a table.  Currently, only the
number of lines cleared are counted.  The multipliers and other
modifications for combos and T-Spins are not yet implemented.

A bunch of menu code was also moved to the second game bank to free up
space on the first bank.  The first bank contains all of the code for
both single and multiplayer games.