Implement changing the time attack mode
The time attack selection is now implemented.  This required changing
the single block selection sprites to operate the same way the active
blocks work in-game.  Ie, the block always uses the same palette index
but the colors in that index change to match the currently active piece.

The piece color tables were also moved to the Utils bank so they are
available to both the game engine and the menus.